[ZEPPELIN-6090] Drop JDK8 from CI
### What is this PR for?
This pull request removes the test with JDK 8. I have kept the JDK matrix to make it easier to switch to the next JDK later.
This pull request also drops python 3.7 & 3.8 tests for the `python`,`rlang` and `zeppelin-jupyter` interpreter. Please note that current python 3.7 and 3.8 are end of life. https://devguide.python.org/versions/

### What type of PR is it?
Improvement

### What is the Jira issue?
 - https://issues.apache.org/jira/browse/ZEPPELIN-6090

### How should this be tested?
* CI

### Questions:
* Does the license files need to update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No


Closes #4875 from Reamer/jdk11_github.

Signed-off-by: Philipp Dallig <philipp.dallig@gmail.com>