[ZEPPELIN-6249] Fix broken Ace Editor theme link in STYLE.md
### What is this PR for?
This PR updates the outdated link to Ace Editor themes in the STYLE.md file.  
The previous link pointed to a non-existing path, and the new link ensures contributors can find the correct theme list.

### What type of PR is it?
Documentation

### Todos
- [x] Update broken theme URL

### What is the Jira issue?
[ZEPPELIN-6249](https://issues.apache.org/jira/browse/ZEPPELIN-6249)

### How should this be tested?
No functional code changes. Just confirm that the updated link is valid and accessible in the rendered markdown.

### Screenshots (if appropriate)
_N/A_

### Questions:
* Does the license files need to update? No  
* Is there breaking changes for older versions? No  
* Does this needs documentation? No


Closes #4992 from hyo-ryeong/feature/ZEPPELIN-6249.

Signed-off-by: Cheng Pan <chengpan@apache.org>