[ZEPPELIN-6282] Add null check condition
### What is this PR for?
This PR was created to prevent an error that occurs in the `InterpreterSetting.getInterpreterGroup` method when `groupId` is null and is retrieved from the Set.
Normally, the groupId parameter does not seem to be passed as null.
I found it when tested service shutdown(#4965).

### What type of PR is it?
Improvement

### Todos
* [x] - Add null check condition

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-6282

### How should this be tested?

### Screenshots (if appropriate)

### Questions:
* Does the license files need to update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Closes #5026 from proceane/feature/ZEPPELIN-6282.

Signed-off-by: ParkGyeongTae <gyeongtae@apache.org>
(cherry picked from commit 078b754ad6abdb74148480f9364920b4a1853085)
Signed-off-by: ParkGyeongTae <gyeongtae@apache.org>