[ZEPPELIN-6265] Support classic Helium viz packages in new UI
### What is this PR for?
This PR enables the use of external Helium visualization packages, which are previously available only in the classic UI, in the new UI.

The process of loading and importing Helium packages, as well as rendering them, closely follows the logic from the classic UI.

Since the new-style visualization classes in the new UI are not compatible with the existing ones, the implementation branches accordingly to handle both.

Classic Helium visualization packages depend on legacy technologies such as AngularJS, jQuery, and several visualization libraries, which had to be included as a result.
For Bootstrap styles, conflicts arose with existing styles, so I copied and modified the original HTML templates to prevent clashes.
If a Helium package attempts to use the original HTML template, it will be replaced with the modified version.

Additionally, I identified that the Helium visualization support classes from the discontinued development of the new UI are no longer in use, so I removed them.

### What type of PR is it?

Improvement

### Todos
* [v] - Added support for loading Helium packages
* [v] - Fixed style mismatches in Helium packages
* [v] - Other functional issues resolved

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-6265

### How should this be tested?
* Enables a classic Helium visualization package in the classic UI, and verify that it works in the new UI.
  - Note: The `/helium` package management page has not yet been added.

### Questions:
* Does the license files need to update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Closes #5006 from tbonelee/helium-vis.

Signed-off-by: ChanHo Lee <chanholee@apache.org>
(cherry picked from commit 948f1652f9277a9e5b25051eee6e0674f137afd3)
Signed-off-by: ChanHo Lee <chanholee@apache.org>